binance_trade/requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
Babel==2.9.1
binance-futures-connector==3.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==40.0.1
dateparser==1.1.7
frozenlist==1.3.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
install==1.3.5
magic-filter==1.0.9
multidict==6.0.4
numpy==1.24.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pyOpenSSL==23.1.1
pyTelegramBotAPI==4.10.0
python-binance==1.0.17
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.28.2
service-identity==21.1.0
six==1.16.0
TA-Lib==0.4.25
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
urllib3==1.26.15
websockets==10.4
yarl==1.8.2
zope.interface==6.0
